boot: Change ":rom%d" boot order rom instance to ":rom%x"
Use hex numbers for the rom instance count in boot order open firmware
device naming.  The ":rom" suffix isn't part of a standard and it's
highly unlikely any rom would have 10 or more drives on it, but this
change makes the code more similar to the numbering of other boot
order devices.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>
Reviewed-by: Markus Armbruster <armbru@redhat.com>