tpm: Refactor hash_log_event BIOS interface function
Refactor the signature of hash_log_event to take individual pointers
as parameters and introduce hash_log_event_int as an function to
be called with the parameters passed from the BIOS interrupt.

Refactor existing callers to hash_log_event that now do not
have to build up the data structures expected by the BIOS interface.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>