tpm: Refactor hash_log_extend_event
Refactor the signature of the hash_log_extend_event to take individual
pointers as parameters and introduce hash_log_extend_event_int as a
function to be called with the parameters passed from the BIOS interrupt.

Refactor existing callers to hash_log_extend_event that now do not
have to build up the data structure expected by the BIOS interface.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>