tpm: Refactor pass_through_to_tpm
Refactor the signature of the pass_through_to_tpm function to take
individual pointers as parameters and introduce pass_through_to_tpm_int
as a function to be called with the parameters passed from the BIOS
interrupt.

Refactor existing callers that now do not have to build up the data
structure expected by the BIOS interface.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>