tpm: Rework TPM interface shutdown support
The 16bit BIOS interface should only shutdown on request from that
interface - errors from the tcp or acpi log setup should not shutdown
the interface.  (Errors from those functions will cause the TPM to be
in a "not working" state which will cause all the 16bit interface
functions to fail.)

Centralize the checking for the interface shutdown condition in
tpm_interrupt_handler32().

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>