tpm: Don't use 16bit BIOS return codes in build_and_send_cmd()
Don't use the return codes from the 16bit BIOS spec in the internal
function build_and_send_cmd().  Instead, return the TIS command status
code of the command or -1 if there was a command transmission failure.
This eliminates the need for a returnCode pointer parameter.

Also, implement debugging dprintf() in build_and_send_cmd() instead of
in every caller.  This replaces the command name with the integer
command id, but it does make the debugging more consistent.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>