pvscsi: fix the comment about lun enumeration
The comment in pvscsi_scan_target (presumably c&p-ed from another
driver) reads that REPORTS LUNS should better be used to enumerate the
luns on the target.

However, according to the Linux driver, the device supports no more than
a single lun per target.

So adjust the comment to tell exactly that.

Signed-off-by: Roman Kagan <rkagan@virtuozzo.com>