ssdt: Fix building of legacy acpi tables on current iasl compiler
Recent versions of the iasl compiler raise an error if the table id is
longer than 8 characters.  Older versions of iasl would silently
truncate the table id to 8 characters.  Change the ssdt-misc and
ssdt-pcihp files to use an 8 character id - this should not directly
impact the generated aml code as the table id was already being
truncated - but may help those wishing to manually compile the tables.

Reported by Michael Tokarev, Vivia Nikolaidou, and several others.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>